requirements.txt
Django==3.0.7
python-decouple==3.3
wordpress-api==1.2.9